requirements.txt
numpy<2
wheel
termcolor
pyyaml
toml
iiif
icalendar
pylint
qrcode ~=7.4
cairosvg
svgutils
#svgutils @ git+https://github.com/btel/svg_utils@1cc5b888f265407d2ea3b1f93c95b2efd29c611c
idna
auditwheel
Cython
qrcode
python-frontmatter
#GeoJSON Requirements
geojson
pyproj
shapely
Pillow
# See https://adamj.eu/tech/2019/03/11/pip-install-from-a-git-repository/
#jxlpy @ git+https://github.com/foosoftsrl/jxlpy@0.9.5
#https://github.com/foosoftsrl/jxlpy/releases/download/0.9.5/jxlpy-0.9.5-cp310-cp310-macosx_11_0_arm64.whl
#jxlpy @ git+https://github.com/olokelo/jxlpy@4cc41c4ec51dc2099c5b4b65235859f869e98ea0
jxlpy>=0.9.5
#pyvips
#PDF tools
pikepdf
fpdf2
# Archiver
beautifulsoup4
httpx
requests
# HDR related
ffmpeg-python
opencv
opencv-contrib-python-headless
docker
pyexiv2
